+++ glean/src/core/pings/maker.ts
@@ -3,9 +3,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { GLEAN_SCHEMA_VERSION, GLEAN_VERSION, PING_INFO_STORAGE, CLIENT_INFO_STORAGE } from "../constants.js";
-import CounterMetricType, { CounterMetric } from "../metrics/types/counter.js";
-import DatetimeMetricType, { DatetimeMetric } from "../metrics/types/datetime.js";
-import { Lifetime } from "../metrics/index.js";
+import CounterMetricType from "../metrics/types/counter.js";
+import { CounterMetric } from "../metrics/types/counter_metric.js";
+import DatetimeMetricType from "../metrics/types/datetime.js";
+import { DatetimeMetric } from "../metrics/types/datetime_metric.js";
 import TimeUnit from "../metrics/time_unit.js";
 import { ClientInfo, PingInfo, PingPayload } from "../pings/ping_payload.js";
 import CommonPingData from "./common_ping_data.js";
@@ -14,6 +15,7 @@ import MetricsDatabase from "../metrics/database.js";
 import EventsDatabase from "../metrics/events_database.js";
 import PingsDatabase from "./database.js";
 import { DebugOptions } from "../debug_options.js";
+import { Lifetime } from "../metrics/lifetime.js";
 
 // The moment the current Glean.js session started.
 const GLEAN_START_TIME = new Date();