+++ glean/tests/core/metrics/database.spec.ts
@@ -5,10 +5,11 @@
 import assert from "assert";
 
 import Database, { isValidInternalMetricsRepresentation } from "../../../src/core/metrics/database";
-import { Lifetime } from "../../../src/core/metrics";
-import StringMetricType, { StringMetric } from "../../../src/core/metrics/types/string";
+import StringMetricType from "../../../src/core/metrics/types/string";
+import { StringMetric } from "../../../src/core/metrics/types/string_metric";
 import { JSONValue } from "../../../src/core/utils";
 import Glean from "../../../src/core/glean";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
 
 describe("MetricsDatabase", function() {
   const testAppId = `gleanjs.test.${this.title}`;