+++ glean/tests/core/metrics/string.spec.ts
@@ -5,8 +5,9 @@
 import assert from "assert";
 
 import Glean from "../../../src/core/glean";
-import StringMetricType, { MAX_LENGTH_VALUE } from "../../../src/core/metrics/types/string";
-import { Lifetime } from "../../../src/core/metrics";
+import StringMetricType from "../../../src/core/metrics/types/string";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
+import { MAX_LENGTH_VALUE } from "../../../src/core/metrics/types/string_metric";
  
 describe("StringMetric", function() {
   const testAppId = `gleanjs.test.${this.title}`;